[refactor] 리뷰어 피드백 반영
- 게임에 사용되는 상수 관리 파일(GameRule.js) 생성
- 게임에 사용되는 조건 숫자 위의 파일에서 import 하는 방식으로 수정
- constants 폴더의 index.js 사용하도록 import 문 수정
- 1라운드만 표기되던 오류 해결을 위해 실행결과와 라운드 표기 함수를 분리
