+++ package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "우아한테크코스 웹 프론트엔드 레벨1 자동차 경주 미션",
   "main": "src/index.js",
+  "type": "module",
   "scripts": {
     "start": "npm run build-prod && node dist/bundle.js",
     "build-dev": "webpack --mode development",
