+++ src/constants/GameRule.js
@@ -1,0 +1,8 @@
+export const MIN_NAME_LENGTH = 1;
+export const MAX_NAME_LENGTH = 5;
+export const MIN_NAME_COUNT = 1;
+export const MIN_CAR_COUNT = 2;
+export const MAX_CAR_COUNT = 5;
+export const MIN_COMMA_COUNT = 1;
+export const MIN_ROUND_COUNT = 1;
+export const MAX_ROUND_COUNT = 5;
