+++ src/constants/index.js
@@ -1,1 +1,2 @@
+export * from './GameRule.js';
 export * from './Message.js';
