+++ src/models/RoundValidator.js
@@ -1,8 +1,8 @@
-import { ERROR_MESSAGE } from '../constants/Message.js';
+import { ERROR_MESSAGE, MIN_ROUND_COUNT, MAX_ROUND_COUNT } from '../constants/index.js';
 
 const RoundValidator = {
   confirmRound(number) {
-    const regex = /^[1-5]$/;
+    const regex = new RegExp(`^[${MIN_ROUND_COUNT}-${MAX_ROUND_COUNT}]$`);
 
     if (!regex.test(number)) {
       throw new Error(ERROR_MESSAGE.round);
